Moved components back to components folder
