+++ pages/services/index.js
@@ -1,5 +1,5 @@
 import React from "react"
-import Services from "./Services"
+import Services from "../../components/Services"
 const index = () => {
     return (
         <div>
